+++ docs/source/api.rst
@@ -1,5 +1,14 @@
 API Reference
 -------------
 
+grill.cook
+~~~~~~~~~~
+
 .. automodule:: grill.cook
+   :members:
+
+grill.usd
+~~~~~~~~~
+
+.. automodule:: grill.usd
    :members:
