+++ grill/cook/__init__.py
@@ -407,81 +407,3 @@ def _find_layer_matching(tokens: typing.Mapping, layers: typing.Iterable[Sdf.Lay
             continue
         return layer
     raise ValueError(f"Could not find layer matching {tokens}. Searched on:\n{pformat(seen)}")
-
-
-@typing.overload
-def edit_context(payload: Sdf.Payload, prim: Usd.Prim) -> Usd.EditContext:
-    """Payload overload"""
-
-
-@typing.overload
-def edit_context(prim: Usd.Prim, query_filter: Usd.PrimCompositionQuery.Filter, target_predicate: typing.Callable) -> Usd.EditContext:
-    """Overload + new def"""
-
-
-@functools.singledispatch
-def edit_context(obj, layer) -> Usd.EditContext:
-    """Doc must be here?"""
-    raise TypeError(f"Not implemented: {locals()}")  # lazy
-
-
-@edit_context.register
-def _(prim: Usd.Prim, query_filter, target_predicate):
-    """Composition arcs target layer stacks. This is a convenience function to
-    get an edit context from a query filter + a filter predicate, using the root layer
-    of the matching target node as the edit target layer.
-    """
-    query = Usd.PrimCompositionQuery(prim)
-    query.filter = query_filter
-    for arc in query.GetCompositionArcs():
-        node = arc.GetTargetNode()
-        if target_predicate(node):
-            target = Usd.EditTarget(node.layerStack.identifier.rootLayer, node)
-            return Usd.EditContext(prim.GetStage(), target)
-    raise ValueError(f"Could not find appropriate node for edit target for {prim} matching {target_predicate}")
-
-
-@edit_context.register
-def _(payload: Sdf.Payload, prim):
-    """Payload"""
-    with Ar.ResolverContextBinder(prim.GetStage().GetPathResolverContext()):
-        # Use Layer.Find since layer should have been open for the prim to exist.
-        layer = Sdf.Layer.Find(payload.assetPath)
-    if not (payload.primPath or layer.defaultPrim):
-        raise ValueError(f"Can't proceed without a prim path to target on payload {payload} for {layer}")
-    path = payload.primPath or layer.GetPrimAtPath(layer.defaultPrim).path
-    logger.debug(f"Searching to target {layer} on {path}")
-
-    def is_valid_target(node):
-        return node.path == path and node.layerStack.identifier.rootLayer == layer
-
-    query_filter = Usd.PrimCompositionQuery.Filter()
-    query_filter.arcTypeFilter = Usd.PrimCompositionQuery.ArcTypeFilter.Payload
-    query_filter.hasSpecsFilter = Usd.PrimCompositionQuery.HasSpecsFilter.HasSpecs
-    return edit_context(prim, query_filter, is_valid_target)
-
-
-@edit_context.register
-def _(variant_set: Usd.VariantSet, layer):
-    """variant set"""
-    with contextlib.suppress(Tf.ErrorException):
-        return variant_set.GetVariantEditContext()
-    # ----- From Pixar -----
-    # pxr.Tf.ErrorException:
-    # 	Error in '...::UsdVariantSet::GetVariantEditTarget' ...: 'Layer <identifier> is not a local layer of stage rooted at layer <identifier>'
-    # https://graphics.pixar.com/usd/docs/api/class_usd_variant_set.html#a83f3adf614736a0b43fa1dd5271a9528
-    # Currently, we require layer to be in the stage's local LayerStack (see UsdStage::HasLocalLayer()), and will issue an error and return an invalid EditTarget if layer is not.
-    # We may relax this restriction in the future, if need arises, but it introduces several complications in specification and behavior.
-    # ---------------------
-    prim = variant_set.GetPrim()
-    name = variant_set.GetName()
-    selection = variant_set.GetVariantSelection()
-    logger.debug(f"Searching target for {prim} with variant {name}, {selection} on {layer}")
-
-    def is_valid_target(node):
-        return node.path.GetVariantSelection() == (name, selection) and layer == node.layerStack.identifier.rootLayer
-
-    query_filter = Usd.PrimCompositionQuery.Filter()
-    query_filter.arcTypeFilter = Usd.PrimCompositionQuery.ArcTypeFilter.Variant
-    query_filter.hasSpecsFilter = Usd.PrimCompositionQuery.HasSpecsFilter.HasSpecs
-    return edit_context(prim, query_filter, is_valid_target)
